% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.filter_na}
\alias{.filter_na}
\title{Remove NA values from a list}
\usage{
.filter_na(l)
}
\arguments{
\item{l}{List to filter}
}
\value{
List with only non-NA elements
}
\description{
Keeps only non-NA elements in a list, used in CSL bibliography generation.
}
\keyword{internal}
