% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.lapply_filter_internal}
\alias{.lapply_filter_internal}
\title{Remove internal metadata from each list element}
\usage{
.lapply_filter_internal(l)
}
\arguments{
\item{l}{List to filter}
}
\value{
List with internal fields removed from each element
}
\description{
Filters out internal fields from each element of a list for CSL export.
}
\keyword{internal}
