% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tables.R
\name{.read_schema_sql}
\alias{.read_schema_sql}
\title{Read SQL schema file}
\usage{
.read_schema_sql(table_name)
}
\arguments{
\item{table_name}{Name of the table (e.g., "persons", "items")}
}
\value{
Character string containing SQL CREATE TABLE statement
}
\description{
Reads a SQL schema file from inst/sql/common/ directory.
}
\keyword{internal}
