% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erudiciondb-package.R
\docType{package}
\name{erudiciondb-package}
\alias{erudiciondb}
\alias{erudiciondb-package}
\title{erudiciondb: Database Configuration and Tools for Storing and Querying Scholars and Scholarship}
\description{
A database schema for scholarship linked to a specific (small) set of scholars along with tools for inserting, updating, querying, and otherwise managing the data along with a reasonable audit trail.
}
\author{
\strong{Maintainer}: Ryan Raaum \email{ryan.raaum@lehman.cuny.edu} (\href{https://orcid.org/0000-0003-4723-5475}{ORCID})

}
\keyword{internal}
