% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{table_columns}
\alias{table_columns}
\title{Get table column names}
\usage{
table_columns(con, table_name)
}
\arguments{
\item{con}{Database connection}

\item{table_name}{Name of table}
}
\value{
Character vector of column names
}
\description{
Wrapper around DBI::dbListFields for test readability.
}
\keyword{internal}
