% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{.items_to_biblio_items}
\alias{.items_to_biblio_items}
\title{A function}
\usage{
.items_to_biblio_items(connection, items)
}
\arguments{
\item{connection}{A pool database connection}

\item{items}{A dataframe of items}
}
\value{
The dataframe with new mean and sum columns
}
\description{
A function
}
