% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{ErudicionDB}
\alias{ErudicionDB}
\title{ErudicionDB Object}
\value{
Nothing
}
\description{
Interface for an ErudicionDB.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{con}}{ErudicionDB. Read-only.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-erudicion_db_object-new}{\code{ErudicionDB$new()}}
\item \href{#method-erudicion_db_object-establish_connection}{\code{ErudicionDB$establish_connection()}}
\item \href{#method-erudicion_db_object-disconnect}{\code{ErudicionDB$disconnect()}}
\item \href{#method-erudicion_db_object-finalize}{\code{ErudicionDB$finalize()}}
\item \href{#method-erudicion_db_object-tbl}{\code{ErudicionDB$tbl()}}
\item \href{#method-erudicion_db_object-add_validator}{\code{ErudicionDB$add_validator()}}
\item \href{#method-erudicion_db_object-add_augmentor}{\code{ErudicionDB$add_augmentor()}}
\item \href{#method-erudicion_db_object-new_object}{\code{ErudicionDB$new_object()}}
\item \href{#method-erudicion_db_object-insert_object}{\code{ErudicionDB$insert_object()}}
\item \href{#method-erudicion_db_object-insert_new_object}{\code{ErudicionDB$insert_new_object()}}
\item \href{#method-erudicion_db_object-retrieve}{\code{ErudicionDB$retrieve()}}
\item \href{#method-erudicion_db_object-find}{\code{ErudicionDB$find()}}
\item \href{#method-erudicion_db_object-match_person}{\code{ErudicionDB$match_person()}}
\item \href{#method-erudicion_db_object-clone}{\code{ErudicionDB$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-erudicion_db_object-new"></a>}}
\if{latex}{\out{\hypertarget{method-erudicion_db_object-new}{}}}
\subsection{Method \code{new()}}{
Insert a new item from an online repository


Create new ErudicionDB object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ErudicionDB$new(dbargs_list)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dbargs_list}}{A list of parameters to create the database}

\item{\code{item_data}}{A list with item and creator entries}

\item{\code{stage}}{Activation stage}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-erudicion_db_object-establish_connection"></a>}}
\if{latex}{\out{\hypertarget{method-erudicion_db_object-establish_connection}{}}}
\subsection{Method \code{establish_connection()}}{
Create ErudicionDB object and connect to a database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ErudicionDB$establish_connection(dbargs_list)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dbargs_list}}{A list of parameters to create the database}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-erudicion_db_object-disconnect"></a>}}
\if{latex}{\out{\hypertarget{method-erudicion_db_object-disconnect}{}}}
\subsection{Method \code{disconnect()}}{
Disconnect from a database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ErudicionDB$disconnect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-erudicion_db_object-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-erudicion_db_object-finalize}{}}}
\subsection{Method \code{finalize()}}{
Clean up after removal of ErudicionDB object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ErudicionDB$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-erudicion_db_object-tbl"></a>}}
\if{latex}{\out{\hypertarget{method-erudicion_db_object-tbl}{}}}
\subsection{Method \code{tbl()}}{
Generate a data frame interface to a database table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ErudicionDB$tbl(tbl_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tbl_name}}{A database table name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-erudicion_db_object-add_validator"></a>}}
\if{latex}{\out{\hypertarget{method-erudicion_db_object-add_validator}{}}}
\subsection{Method \code{add_validator()}}{
Add a function to validate a new database table entry
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ErudicionDB$add_validator(for_what, validate_function)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{for_what}}{The singular name of the table}

\item{\code{validate_function}}{A function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-erudicion_db_object-add_augmentor"></a>}}
\if{latex}{\out{\hypertarget{method-erudicion_db_object-add_augmentor}{}}}
\subsection{Method \code{add_augmentor()}}{
Add a function to augment a new database table entry
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ErudicionDB$add_augmentor(for_what, augment_function)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{for_what}}{The singular name of the table}

\item{\code{augment_function}}{A function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-erudicion_db_object-new_object"></a>}}
\if{latex}{\out{\hypertarget{method-erudicion_db_object-new_object}{}}}
\subsection{Method \code{new_object()}}{
Fully create, augment, and validate object for insertion into database.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ErudicionDB$new_object(object_type, object, stage = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object_type}}{Kind of object (singular of table name)}

\item{\code{object}}{List of object data}

\item{\code{stage}}{Activation stage (default: 0=Active)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-erudicion_db_object-insert_object"></a>}}
\if{latex}{\out{\hypertarget{method-erudicion_db_object-insert_object}{}}}
\subsection{Method \code{insert_object()}}{
Insert a properly created, augmented, and validated object into the db.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ErudicionDB$insert_object(object)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object}}{List with proper object for database.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-erudicion_db_object-insert_new_object"></a>}}
\if{latex}{\out{\hypertarget{method-erudicion_db_object-insert_new_object}{}}}
\subsection{Method \code{insert_new_object()}}{
Properly create and insert an object into db in one call.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ErudicionDB$insert_new_object(object_type, object, stage = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object_type}}{Kind of object (singular of table name)}

\item{\code{object}}{List of object data}

\item{\code{stage}}{Activation stage (default: 0=Active)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-erudicion_db_object-retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-erudicion_db_object-retrieve}{}}}
\subsection{Method \code{retrieve()}}{
Retrieve objects from the database.

Note that submitting a vector of object_id's will return multiple
objects. Also note that the search can use a different column using
the \code{by} parameter, in which case the \code{object_id} parameter is the
search term(s) for that other column.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ErudicionDB$retrieve(
  object_type,
  object_id,
  stage = 0,
  revision = "max",
  by = NULL,
  as_list = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object_type}}{Kind of object (singular of table name)}

\item{\code{object_id}}{The object_id in the database}

\item{\code{stage}}{Activation stage (default: 0=Active)}

\item{\code{revision}}{Revision of the object (default: "max"=latest)}

\item{\code{by}}{Retrieve by some column other than object_id (default: NULL=no)}

\item{\code{as_list}}{Return as list (default: TRUE; FALSE = as data frame)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-erudicion_db_object-find"></a>}}
\if{latex}{\out{\hypertarget{method-erudicion_db_object-find}{}}}
\subsection{Method \code{find()}}{
Find an object in the database.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ErudicionDB$find(object_type, object_data, stage = 0, revision = "max")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object_type}}{Kind of object (singular of table name)}

\item{\code{object_data}}{An object in list form}

\item{\code{stage}}{Activation stage (default: 0=Active)}

\item{\code{revision}}{Revision of the object (default: "max"=latest)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-erudicion_db_object-match_person"></a>}}
\if{latex}{\out{\hypertarget{method-erudicion_db_object-match_person}{}}}
\subsection{Method \code{match_person()}}{
Find a focal person from an item_person.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ErudicionDB$match_person(person_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{person_data}}{An item_person in list form}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-erudicion_db_object-clone"></a>}}
\if{latex}{\out{\hypertarget{method-erudicion_db_object-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ErudicionDB$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
